Major major changes. Added a blocking API to ZmdBackend. Fixed some stuff, including formatting. My log messages rock don't they...

+++ trunk/kzmdupdater/kzmdupdater/UpdaterCore.h
@@ -41,7 +41,7 @@ struct Progress {
 	QValueList<QString> messages;
 	int expectedTime;
 	int remainingTime;
-	float percent;
+	double percent;
 };
 
 struct Service {
